Hash fct

+++ includes/hashmap.h
@@ -6,7 +6,7 @@
 /*   By: nrechati <nrechati@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/21 15:07:32 by nrechati          #+#    #+#             */
-/*   Updated: 2019/03/21 17:27:03 by nrechati         ###   ########.fr       */
+/*   Updated: 2019/03/21 17:52:24 by nrechati         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@
 
 typedef struct	s_hash
 {
-	int		map_size;
-	int		used;
+	size_t		map_size;
+	size_t		used;
 	t_list	**map;
 }				t_hash;
 
@@ -29,6 +29,6 @@ typedef struct	s_hnode
 }				t_hnode;
 
 uint32_t	ft_hash_str(char *key);
-int			ft_norm_hash(uint32_t hash, int tab_size);
+uint32_t	ft_norm_hash(uint32_t hash, size_t tab_size);
 
 #endif
